+++ BusinessObjects/Dtos/Orders/OrderResponse.cs
@@ -15,7 +15,7 @@ namespace BusinessObjects.Dtos.Orders
     {
         public Guid OrderId { get; set; }
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public string OrderCode { get; set; }
         public DateTime CreatedDate { get; set; }
         public PaymentMethod PaymentMethod { get; set; }
@@ -40,7 +40,7 @@ namespace BusinessObjects.Dtos.Orders
     {
         public Guid OrderId { get; set; }
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public string OrderCode { get; set; }
         public DateTime CreatedDate { get; set; }
         public PaymentMethod PaymentMethod { get; set; }
