+++ BusinessObjects/Entities/Auction.cs
@@ -13,7 +13,7 @@ public class Auction
     public int DepositFee { get; set; }
     public Shop Shop { get; set; }
     public Guid ShopId { get; set; }
-    public int StepIncrement { get; set; }
+    public decimal StepIncrement { get; set; }
     public AuctionFashionItem AuctionFashionItem { get; set; }
     public Guid AuctionFashionItemId { get; set; }
     public AuctionStatus Status { get; set; }
